Expand looping range for new bih options
The new options for biharmonic viscosity require additional points of
Ah, as a part of del2u and del2v.

* New index limit parameters are added, along with a changable
additional stencil parameter, so that then stencil in the old scheme
does not have to expand.
* The relevant do-loop ranges are now replaced by the new sets of
parameters. The part for calculating Leith may need some re-examination.
It was not tested.
* Comments added before each major if-block to give a clearer view of
the genernal order of the calculation.