Separate SAL from tidal_forcing
Calculation for self-attraction and loading (SAL) is separated as a new
module (MOM_self_attr_load) from module MOM_tidal_forcing, as SAL is a
process not limited to tides.

The new module includes both online spherical harmonics method and
scalar approximation. Read-in method (TIDAL_SAL_FROM_FILE) is kept in
the tidal forcing module as it is specific to tides. For the iterative
method, the updating part is in the new SAL model, while the read-in
part remains in the tidal_forcing module.

* A new parameter CALCULATE_SAL is added, which controls SAL
calculation in PressureForce independent of whether tides is on or not.
The default of CALCULATE_SAL is TIDES to avoid making changes in old
MOM_input.
* For SAL related parameters, their names still incorrectly contain
TIDE or TIDAL. This will be addressed in the following commits.
* A new diagnostic is added in PressureForce to output calculated SAL
fields. Note that the 'e_tidal' diagnostic is unchanged and
still includes SAL field for backward compatibility.