Recover old answers with tides in Boussinesq mode
This commit addresses the issue of bit level answer change with tides
in Boussinesq mode.

* A runtime parameter TIDES_ANSWER_DATE is added to restore old answers
before the SAL module is added. The answer change is due to an reorder
of summations of the SAL and tidal forcing terms.
* The new version flag only applies to the analytical pressure force in
Boussinesq mode, which is virtually the only configuration widely used
and included in the test suite.
* For Montgomery pressure and non-Boussinesq mode, the code is
refactored in a more readable way.