Modify tidal forcing calculation in Boussinesq
Previously, tidal geopotential height was added as a correction to
interface height. While this method is computationally efficient, it is
not strictly correct in Boussinesq mode, as interface height corrected
by tides is not accurate for calculating pressure.

A date integer is added for TIDES_ANSWER_DATE. Values below 20240901
recovers the old answer. The new date integer does not affect
non-Boussinesq mode.