Rename variables with tides in Pressure Force
SSH is renamed to pbot_anom. Tides related variable names are also
changed in order to be less confusing. Notably, e_sal_tide is renamed
to e_sal_and_tide (the summation of sal and tide), not to be confused
with e_tide_sal, which is renamed to e_tidal_sal (sal from tides).