Remove rescaling bottom pressure
Instead of rescaling bottom pressure to height unit, calc_Loving_scaling
 is modified to be conditionally dimensional. When calculating
self-attraction and loading, Love numbers are now dimensional when
bottom pressure anomaly is used as an input. This change eliminate Love
numbers' dependence on mean seawater density.

A new coefficient called linear_scaling is added to SAL CS for the same
purpose, although to use bottom pressure anomaly for scalar
approximation is not quite justifiable. A WARNING is given when users
try to do that.