Fix bug with hatvel in cont solver
When hatvel is called at corrector steps, BT_cont is not passed to cont
solver, this creates an issue with zonal/meridional_flux_thickness as
BT_cont%h_[uv] is not present. This commit changes h_[uv] in
zonal/meridional_flux_thickness to optional and uses a separate call
for calculating hatvel.

Note that the algorithm is very redundant and needs to be optimized
later on.