Porous medium rework
This commit includes reworks on two aspects of porous medium.

1. Distinguish `dz` (interface distance) and the stretched thickness
(equivalence to pressure change)
* Add a new variable `h_prime` as the stretched thickness, which has the
same unit as `h`. Recover `dz`, which has the unit of m.
* `h_prime` is used as the input pressure force.
* Remove `dz` input from step_MOM_dyn_*. Instead, local variables are
used for `h_prime` and `dz`
* `thickness_to_dz` calls for `dz` are recovered when porous medium is
off.
* Add `use_pormed` flag to `MOM_dyn_*` CS and use it to control wheather
`h_prime` is calculated or the same of `h`.

2. Rewrite conversion functions in `MOM_interface_heights`.
* Two subroutines (`height_from_vol_monomial` and
`vol_from_height_monomial`) to convert between height and volume below
using the monomial profile. This fixes a bug in the previous version.
`vol_from_height_monomial` works the same as a subroutine in
`MOM_porous_barriers` and needs to be merged later on.
* Subroutine `adjust_h_subgrid_topo` is primarily used in initializating
new runs. Given `dz`, the actual volume/mass `h` is calculated. Rather
than a standardalone function, the subroutine is used as a call within
a number of initialization subroutines incluing `dz_to_thickness`
* Subroutine `thickness_to_dz_subgrid_topo` calculate `h_prime` from `h`
and `dz`. This implementation utilizes `dz` calculated from
`thickness_to_dz`.

Other minor changes

* Change the signs of depc_low/ave/hgh, so that they are now consistent
with bathyT.
* Temporarily remove the derived fields m1 and m2