Control SAL calculation frequency
This commit adds the option to increase the time interval of updating
self-attraction and loading (SAL) term beyond every (baroclinic) dynamic
timestep.

* A new runtime parameter `SAL_RECALC_PERIOD` is added and has the unit
of seconds. Updating SAL is orchestrated by each MOM_dynamics_*.

* Local field `e_sal` to an attribute of PressureForce_FV_CS type, which
allows it to be saved and reused if SAL is not computed at every
timestep.

* Accordingly, PressureForce_FV_end and PressureForce_end subroutines
are reinstated.

* Interface of subroutine PressureForce() now includes an optional input
`update_SAL`

* The corresponding changes in Montgomery PGF are not yet added.