Disable adjustment of BT mass source in corrector step
An adjustment to the barotropic mass source used in the correction step
of RK2 was making matters worse for layers under a grounded ice shelf.
This adjustment essentially does not do anything for the deep ocean
(i.e. was tiny tiny) so we suspect this was a hold over from the earlier
BT algorithms.